Add read-through cache for opslevel registrations.
The idea here is that if the parsed registration doesn't change, we don't
need to do any work. The cache has an expiry equal to the `resync` period
that is defined. So, if the parsed registration doesn't change, at most
we'll hard-resync once per resync interval.

This was added to address the 1000s of `ServiceGet` requests that we get
per day per service, without corresponding ServiceUpdate calls. This means
that while the kubernetes kind is updated, the information we parse from it
does not change.